services/api/requirements.txt
haystack-ai
ollama-haystack
elasticsearch-haystack
huggingface_hub
flask
flask_limiter
python-dotenv
gunicorn